alpha: reverting alpha value back to float and casting to double instead ref. #13

+++ slock.c
@@ -600,7 +600,7 @@ lockscreen(Display *dpy, struct xrandr *rr, int screen)
 			drawlogo(dpy, lock, INIT);
 			#endif // DWM_LOGO_PATCH
 			#if ALPHA_PATCH
-			unsigned int opacity = (unsigned int)(alpha * 0xffffffff);
+			unsigned int opacity = (unsigned int)((double)alpha * 0xffffffff);
 			XChangeProperty(dpy, lock->win, XInternAtom(dpy, "_NET_WM_WINDOW_OPACITY", False), XA_CARDINAL, 32, PropModeReplace, (unsigned char *)&opacity, 1L);
 			XSync(dpy, False);
 			#endif // ALPHA_PATCH
